Add CreateInterest endpoint to InterestController
This commit introduces a new HTTP POST endpoint in the
`InterestController` class. The `CreateInterest` method
allows clients to create a new interest by accepting an
`Interest` object from the query parameters. It adds the
interest to the repository, saves the changes, and returns
a `201 Created` response with the details of the newly
created interest.